완성

+++ main_window.py
@@ -10,6 +10,7 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 from Download_Manager import *
 import FilePDFManager
 import threading
+import os
 
 class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
@@ -42,12 +43,20 @@ class Ui_MainWindow(object):
         self.pushButton_Search = QtWidgets.QPushButton(self.centralwidget)
         self.pushButton_Search.setGeometry(QtCore.QRect(380, 10, 121, 81))
         self.pushButton_Search.setObjectName("pushButton_Search")
-        self.lineEdit_FileSearch = QtWidgets.QLineEdit(self.centralwidget)
-        self.lineEdit_FileSearch.setGeometry(QtCore.QRect(120, 120, 170, 20))
-        self.lineEdit_FileSearch.setObjectName("lineEdit_FileSearch")
+        #self.lineEdit_FileSearch = QtWidgets.QLineEdit(self.centralwidget)
+        #self.lineEdit_FileSearch.setGeometry(QtCore.QRect(120, 120, 170, 20))
+        #self.lineEdit_FileSearch.setObjectName("lineEdit_FileSearch")
         self.label_3 = QtWidgets.QLabel(self.centralwidget)
         self.label_3.setGeometry(QtCore.QRect(10, 120, 121, 20))
         self.label_3.setObjectName("label_3")
+        #self.pushButton_Log
+        self.pushButton_Log = QtWidgets.QPushButton(self.centralwidget)
+        self.pushButton_Log.setGeometry(QtCore.QRect(425, 120, 75, 23))
+        self.pushButton_Log.setObjectName('pushButton_Log')
+        #self.pushButton_Refresh
+        self.pushButton_Refresh = QtWidgets.QPushButton(self.centralwidget)
+        self.pushButton_Refresh.setGeometry(QtCore.QRect(345, 120, 75, 23))
+        self.pushButton_Refresh.setObjectName('pushButton_Refresh')
         self.label_progress = QtWidgets.QLabel(self.centralwidget)
         self.label_progress.setGeometry(QtCore.QRect(290, 20, 56, 12))
         self.label_progress.setText("")
@@ -105,13 +114,13 @@ class Ui_MainWindow(object):
 
     def retranslateUi(self, MainWindow):
         _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
+        MainWindow.setWindowTitle(_translate("MainWindow", "GoogleSearch"))
         self.label_page.setText(_translate("MainWindow", "페이지 수 : "))
         self.label.setText(_translate("MainWindow", "검색어"))
         self.label_2.setText(_translate("MainWindow", "파일 위치"))
         self.toolButton_Path.setText(_translate("MainWindow", "..."))
         self.pushButton_Search.setText(_translate("MainWindow", "검색\n""\n""다운로드"))
-        self.label_3.setText(_translate("MainWindow", "파일 목록에서 검색"))
+        self.label_3.setText(_translate("MainWindow", "파일 목록"))
         self.label_4.setText(_translate("MainWindow", "추출"))
         self.pushButton_Extract.setText(_translate("MainWindow", "추출"))
         self.pushButton__Remove.setText(_translate("MainWindow", "삭제"))
@@ -119,6 +128,8 @@ class Ui_MainWindow(object):
         self.label_Extract_Line.setText(_translate("MainWindow", "줄 수"))
         self.pushButton_ExtractnOpen.setText(_translate("MainWindow", "추출 후 열기"))
         self.lineEdit_Extract_Line.setText(_translate("MainWindow", '1'))
+        self.pushButton_Log.setText(_translate("MainWindow", "검색 기록"))
+        self.pushButton_Refresh.setText(_translate("MainWindow", "새로고침"))
 
         ######
         #do
@@ -128,9 +139,25 @@ class Ui_MainWindow(object):
         self.pushButton_OpenFile.clicked.connect(self.openbtn_click)
         self.pushButton__Remove.clicked.connect(self.deletebtn_click)
         self.pushButton_Extract.clicked.connect(self.extractbtn_click)
-
+        self.pushButton_ExtractnOpen.clicked.connect(self.extractnopenbtn_click)
+        self.pushButton_Log.clicked.connect(self.OpenLog)
+        self.pushButton_Refresh.clicked.connect(self.refresh_list)
+        #self.lineEdit_FileSearch.textChanged.connect(self.filter_changed)
         self.Running = False
         self.Stop = False
+
+    #def filter_changed(self):
+    #    self.filter = self.lineEdit_FileSearch.text()
+    #    self.refresh_list()
+
+    def OpenLog(self):
+        f = open('log.txt', 'a')
+        f.close()
+
+        path = os.path.dirname(os.path.abspath(__file__)).replace('/', '\\')
+        t = threading.Thread(target=FilePDFManager.OpenFile,
+                             args=(path, 'log.txt'))
+        t.start()
 
     def searchbtn_click(self):
 
@@ -147,9 +174,13 @@ class Ui_MainWindow(object):
         elif self.lineEdit_Path.text() == '':
             QtWidgets.QMessageBox.about(MainWindow, 'Error', '폴더를 선택해주세요')
 
+        elif self.lineEdit_page.text() == '':
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '빈 칸을 채워주세요')
+
         else:
             if self.Running == False:
                 self.Running = True
+                self.page = int(self.lineEdit_page.text())
                 self.label_progress.setText('Downloading...')
                 t = threading.Thread(target=self.download)
                 t.daemon = True
@@ -159,7 +190,7 @@ class Ui_MainWindow(object):
 
         path = self.lineEdit_Path.text().replace('/', '\\')
         keyword = self.lineEdit_Search.text()
-        self.download_manager = Download_Manager(keyword=keyword, path=path)
+        self.download_manager = Download_Manager(keyword=keyword, path=path, page_limit=self.page)
 
         while (1):
             self.refresh_list()
@@ -173,6 +204,7 @@ class Ui_MainWindow(object):
                 self.download_manager.page.DriverQuit()
                 self.Running = False
                 self.label_progress.setText('Done!')
+                self.refresh_list()
                 break
 
     def refresh_list(self):
@@ -187,12 +219,20 @@ class Ui_MainWindow(object):
         self.lineEdit_Path.setText(fname)
         self.path = fname.replace('/', '\\')
 
-        self.listView.setModel(self.model)
-        self.listView.setRootIndex(self.model.index(self.path))
+        self.refresh_list()
 
     def deletebtn_click(self):
 
-        filename = self.listView.currentIndex().data()
+        try:
+            filename = self.listView.currentIndex().data()
+        except:
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
+        if(filename == None):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
         FilePDFManager.DeleteFile(self.path, filename)
         self.refresh_list()
 
@@ -202,16 +242,62 @@ class Ui_MainWindow(object):
             QtWidgets.QMessageBox.about(MainWindow, 'Error', '빈 칸을 채워주세요')
             return False
 
-        filename = self.listView.currentIndex().data()
+        elif(self.lineEdit_Extract_Line.text == ''):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '빈 칸을 채워주세요')
+            return False
+        try:
+            filename = self.listView.currentIndex().data()
+        except:
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
+        if (filename == None):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
         line = int(self.lineEdit_Extract_Line.text())
         data = FilePDFManager.Extractor(self.path, filename,
                                         self.lineEdit_Extract.text(), line)
         filename = filename.split('.')[0]
         FilePDFManager.mktext(self.path, filename, data)
 
+    def extractnopenbtn_click(self):
+
+        if (self.lineEdit_Extract.text() == ''):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '빈 칸을 채워주세요')
+            return False
+
+        elif (self.lineEdit_Extract_Line.text == ''):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '빈 칸을 채워주세요')
+            return False
+        try:
+            filename = self.listView.currentIndex().data()
+        except:
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
+        if (filename == None):
+            QtWidgets.QMessageBox.about(MainWindow, 'Error', '파일을 선택해주세요')
+            return False
+
+        line = int(self.lineEdit_Extract_Line.text())
+        data = FilePDFManager.Extractor(self.path, filename,
+                                        self.lineEdit_Extract.text(), line)
+        filename = filename.split('.')[0]
+        FilePDFManager.mktext(self.path, filename, data)
+
+        t = threading.Thread(target=FilePDFManager.OpenFile,
+                             args=(self.path, filename))
+        t.start()
+
     def openbtn_click(self):
 
+        filename = self.listView.currentIndex().data()
         print(self.listView.currentIndex().data())
+        t = threading.Thread(target=FilePDFManager.OpenFile,
+                             args=(self.path, filename))
+        t.start()
+        #FilePDFManager.OpenFile(self.path, filename)
 
     def __del__(self):
         self.download_manager.page.DriverQuit()
